requirements.txt
selenium==4.11.2
httpx==0.26.0
bs4==0.0.1
python-dotenv==1.0.0
webdriver-manager==4.0.0
pytz==2023.3.post1